requirements.txt
AfricastalkingGateway==1.8
amqp==2.2.2
appdirs==1.4.3
beautifulsoup4==4.5.3
billiard==3.5.0.3
celery==4.1.0
certifi==2017.4.17
chardet==3.0.4
dj-database-url==0.4.1
Django==1.10.4
django-anymail==0.10
django-autoslug==1.9.3
django-debug-toolbar==1.8
django-discover-runner==1.0
django-multiselectfield==0.1.7
django-pipeline==1.6.13
django-redis==4.8.0
django-role-permissions==2.0.1
django-taggit==0.22.1
django-webtest==1.9.1
djangorestframework==3.7.3
futures==3.1.1
gunicorn==19.6.0
idna==2.5
kombu==4.1.0
Markdown==2.6.8
olefile==0.44
packaging==16.8
Pillow==4.2.1
psycopg2==2.6.2
pyparsing==2.2.0
python-decouple==3.0
python-memcached==1.58
pytz==2017.2
redis==2.10.6
requests==2.18.1
six==1.10.0
sqlparse==0.2.3
urllib3==1.21.1
vine==1.1.4
waitress==1.0.2
WebOb==1.7.2
WebTest==2.0.27
whitenoise==3.2
